Separated network server and network client logic.
