+++ V2UnityDiscordIntercept/Patches/DiscordController_DisconnectNetwork2_Patch.cs
@@ -15,8 +15,8 @@ namespace V2UnityDiscordIntercept.Patches
 
             Demo.instance.LeaveLobby(0L);
 
-            if (Plugin.Network.IsServer)
-                Plugin.Network.DeleteLobby();
+            if (DiscordController.IsOwner())
+                Plugin.Server.DeleteLobby();
 
             return false;
         }
